micros-pipeline/requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dependency-injector==4.41.0
dnspython==2.4.2
frozenlist==1.4.0
grpcio==1.60.0
grpcio-reflection==1.60.0
grpcio-tools==1.60.0
idna==3.4
loguru==0.7.2
motor==3.3.1
multidict==6.0.4
numpy==1.26.1
pandas==2.1.2
protobuf==4.25.1
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.5.0
PyMySQL==1.1.0
python-consul2==0.1.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.3
varname==0.13.0
tqdm==4.66.2
nest-asyncio==1.6.0
